partly added property view with controller, model etc. Images returned properly from databse needs more work to have each correct image for each property only returned

+++ fuel/core/classes/image/imagemagick.php
@@ -1,15 +1,13 @@
 <?php
-
 /**
  * Part of the Fuel framework.
  *
- * Image manipulation class.
- *
- * @package		Fuel
- * @version		1.0
- * @license		MIT License
- * @copyright	2010 - 2011 Fuel Development Team
- * @link		http://fuelphp.com
+ * @package    Fuel
+ * @version    1.7
+ * @author     Fuel Development Team
+ * @license    MIT License
+ * @copyright  2010 - 2014 Fuel Development Team
+ * @link       http://fuelphp.com
  */
 
 namespace Fuel\Core;
@@ -19,12 +17,12 @@ class Image_Imagemagick extends \Image_Driver
 
 	protected $image_temp = null;
 	protected $accepted_extensions = array('png', 'gif', 'jpg', 'jpeg');
-	protected $size_cache = null;
+	protected $sizes_cache = null;
 	protected $im_path = null;
 
-	public function load($filename, $return_data = false)
+	public function load($filename, $return_data = false, $force_extension = false)
 	{
-		extract(parent::load($filename));
+		extract(parent::load($filename, $return_data, $force_extension));
 
 		$this->clear_sizes();
 		if (empty($this->image_temp))
@@ -33,15 +31,15 @@ class Image_Imagemagick extends \Image_Driver
 			{
 				$this->image_temp = $this->config['temp_dir'].substr($this->config['temp_append'].md5(time() * microtime()), 0, 32).'.png';
 			}
-			while (file_exists($this->image_temp));
+			while (is_file($this->image_temp));
 		}
-		elseif (file_exists($this->image_temp))
+		elseif (is_file($this->image_temp))
 		{
 			$this->debug('Removing previous temporary image.');
 			unlink($this->image_temp);
 		}
 		$this->debug('Temp file: '.$this->image_temp);
-		if (!file_exists($this->config['temp_dir']) || !is_dir($this->config['temp_dir']))
+		if ( ! is_dir($this->config['temp_dir']))
 		{
 			throw new \RuntimeException("The temp directory that was given does not exist.");
 		}
@@ -85,20 +83,46 @@ class Image_Imagemagick extends \Image_Driver
 		$this->clear_sizes();
 	}
 
+	protected function _flip($direction)
+	{
+		switch ($direction)
+		{
+			case 'vertical':
+			$arg = '-flip';
+			break;
+
+			case 'horizontal':
+			$arg = '-flop';
+			break;
+
+			case 'both':
+			$arg = '-flip -flop';
+			break;
+
+			default: return false;
+		}
+		$image = '"'.$this->image_temp.'"';
+		$this->exec('convert', $image.' '.$arg.' '.$image);
+	}
+
 	protected function _watermark($filename, $position, $padding = 5)
 	{
-		extract(parent::_watermark($filename, $x, $y));
+		$values = parent::_watermark($filename, $position, $padding);
+		if ($values == false)
+		{
+			throw new \InvalidArgumentException("Watermark image not found or invalid filetype.");
+		}
 
-		$image = '"'.$this->image_temp.'"';
-		$filename = '"'.$filename.'"';
+		extract($values);
 		$x >= 0 and $x = '+'.$x;
 		$y >= 0 and $y = '+'.$y;
 
+		$image = '"'.$this->image_temp.'"';
 		$this->exec(
 			'composite',
 			'-compose atop -geometry '.$x.$y.' '.
 			'-dissolve '.$this->config['watermark_alpha'].'% '.
-			$filename.' '.$image.' '.$image
+			'"'.$filename.'" "'.$this->image_temp.'" '.$image
 		);
 	}
 
@@ -137,12 +161,12 @@ class Image_Imagemagick extends \Image_Driver
 
 		$image = '"'.$this->image_temp.'"';
 		$r = $radius;
-		$command = $image." ( +clone -alpha extract ".
+		$command = $image." \\( +clone -alpha extract ".
 			( ! $tr ? '' : "-draw \"fill black polygon 0,0 0,$r $r,0 fill white circle $r,$r $r,0\" ")."-flip ".
 			( ! $br ? '' : "-draw \"fill black polygon 0,0 0,$r $r,0 fill white circle $r,$r $r,0\" ")."-flop ".
 			( ! $bl ? '' : "-draw \"fill black polygon 0,0 0,$r $r,0 fill white circle $r,$r $r,0\" ")."-flip ".
 			( ! $tl ? '' : "-draw \"fill black polygon 0,0 0,$r $r,0 fill white circle $r,$r $r,0\" ").
-			') -alpha off -compose CopyOpacity -composite '.$image;
+			'\\) -alpha off -compose CopyOpacity -composite '.$image;
 		$this->exec('convert', $command);
 	}
 
@@ -157,7 +181,7 @@ class Image_Imagemagick extends \Image_Driver
 		$is_loaded_file = $filename == null;
 		if ( ! $is_loaded_file or $this->sizes_cache == null or !$usecache)
 		{
-			$reason = ($filename != null ? "filename" : ($this->size_cache == null ? 'cache' : 'option'));
+			$reason = ($filename != null ? "filename" : ($this->sizes_cache == null ? 'cache' : 'option'));
 			$this->debug("Generating size of image... (triggered by $reason)");
 
 			if ($is_loaded_file and ! empty($this->image_temp))
@@ -165,7 +189,7 @@ class Image_Imagemagick extends \Image_Driver
 				$filename = $this->image_temp;
 			}
 
-			$output = $this->exec('identify', '-format "%[fx:w] %[fx:h]" "'.$filename.'"[0]');
+			$output = $this->exec('identify', '-format "%w %h" "'.$filename.'"[0]');
 			list($width, $height) = explode(" ", $output[0]);
 			$return = (object) array(
 				'width' => $width,
@@ -186,21 +210,21 @@ class Image_Imagemagick extends \Image_Driver
 		return $return;
 	}
 
-	public function save($filename, $permissions = null)
+	public function save($filename = null, $permissions = null)
 	{
 		extract(parent::save($filename, $permissions));
 
 		$this->run_queue();
 		$this->add_background();
-		
+
 		$filetype = $this->image_extension;
 		$old = '"'.$this->image_temp.'"';
 		$new = '"'.$filename.'"';
-		
+
 		if(($filetype == 'jpeg' or $filetype == 'jpg') and $this->config['quality'] != 100)
 		{
 			$quality = '"'.$this->config['quality'].'%"';
-			$this->exec('convert', $old.' -quality '.$quality.' '.strtolower($filetype).' '.$new);
+			$this->exec('convert', $old.' -quality '.$quality.' '.$new);
 		}
 		else
 		{
@@ -221,9 +245,9 @@ class Image_Imagemagick extends \Image_Driver
 
 		$this->run_queue();
 		$this->add_background();
-		
+
 		$image = '"'.$this->image_temp.'"';
-		
+
 		if(($filetype == 'jpeg' or $filetype == 'jpg') and $this->config['quality'] != 100)
 		{
 			$quality = '"'.$this->config['quality'].'%"';
@@ -282,7 +306,7 @@ class Image_Imagemagick extends \Image_Driver
 	 * @param   boolean  $passthru  Returns the output if false or pass it to browser.
 	 * @return  mixed    Either returns the output or returns nothing.
 	 */
-	private function exec($program, $params, $passthru = false)
+	protected function exec($program, $params, $passthru = false)
 	{
 		//  Determine the path
 		$this->im_path = realpath($this->config['imagemagick_dir'].$program);
@@ -319,13 +343,16 @@ class Image_Imagemagick extends \Image_Driver
 	 */
 	protected function create_color($hex, $alpha)
 	{
-		export($this->create_hex_color($hex));
+		extract($this->create_hex_color($hex));
 		return "\"rgba(".$red.", ".$green.", ".$blue.", ".round($alpha / 100, 2).")\"";
 	}
 
 	public function __destruct()
 	{
-		unlink($this->image_temp);
+		if (is_file($this->image_temp))
+		{
+			unlink($this->image_temp);
+		}
 	}
 }
